Correct standardization, correct mean values replacement and add offset
For standarsization: standarsize the test set according to the mean and the std of the train set. Apply the standardize column by column. For the mean values replacement : replace the test missing values with the mean of the training set. For the offset : add a column of 1 in the features.